analytics-service/requirements.txt
fastapi
uvicorn
asyncpg
redis
python-dotenv
ua-parser
requests